requirements.txt
llama-index
openai
pypdf
python-dotenv
streamlit
langchain
langchain-community